requirements.txt
numpy==1.25.2
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pandas==2.2.3
protobuf==3.20.3
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.14.1
tqdm==4.67.0
ultralytics==8.3.32
scikit-image==0.24.0
filterpy==1.4.5
